Crate forest_vm

Macros

actor_error
Convenience macro for generating Actor Errors

Structs

ActorError
The error type that gets returned by actor method calls.

ActorState
State of all actor implementations.

EMPTY_ARR_BYTES
Cbor bytes of an empty array serialized.

EMPTY_ARR_CID
Cid of the empty array Cbor bytes (EMPTY_ARR_BYTES).

Serialized
Serialized bytes to be used as parameters into actor methods. This data is (de)serialized as a byte string.

Enums

ExitCode
ExitCode defines the exit code from the VM execution.

Constants

METHOD_CONSTRUCTOR
Base actor constructor method.

METHOD_SEND
Base actor send method.

Type Definitions

DealID
Deal identifier used in market and miner actors

MethodNum
Method number indicator for calling actor methods.

TokenAmount
Token type to be used within the VM.