Enum forest_vm::ExitCode

#[repr(u64)]
pub enum ExitCode {

    Ok,
    SysErrSenderInvalid,
    SysErrSenderStateInvalid,
    SysErrInvalidMethod,
    SysErrActorPanic,
    SysErrInvalidReceiver,
    SysErrInsufficientFunds,
    SysErrOutOfGas,
    SysErrForbidden,
    SysErrIllegalActor,
    SysErrIllegalArgument,
    SysErrReserved2,
    SysErrReserved3,
    SysErrReserved4,
    SysErrReserved5,
    SysErrReserved6,
    ErrIllegalArgument,
    ErrNotFound,
    ErrForbidden,
    ErrInsufficientFunds,
    ErrIllegalState,
    ErrSerialization,
    ErrTooManyProveCommits,
    ErrPlaceholder,

}

ExitCode defines the exit code from the VM execution.

Variants

Ok

SysErrSenderInvalid

Indicates failure to find an actor in the state tree.

SysErrSenderStateInvalid

Indicates failure to find the code for an actor.

SysErrInvalidMethod

Indicates failure to find a method in an actor.

SysErrActorPanic

Used for catching panics currently. (marked as unused/SysErrReserved1 in go impl though)

SysErrInvalidReceiver

Indicates a message sender has insufficient funds for a message’s execution.

SysErrInsufficientFunds

Indicates a message invocation out of sequence.

SysErrOutOfGas

Indicates message execution (including subcalls) used more gas than the specified limit.

SysErrForbidden

Indicates a message execution is forbidden for the caller.

SysErrIllegalActor

Indicates actor code performed a disallowed operation. Disallowed operations include:

mutating state outside of a state acquisition block

failing to invoke caller validation

aborting with a reserved exit code (including success or a system error).

SysErrIllegalArgument

Indicates an invalid argument passed to a runtime method.

SysErrReserved2

Reserved exit codes, do not use.

SysErrReserved3

SysErrReserved4

SysErrReserved5

SysErrReserved6

ErrIllegalArgument

Indicates a method parameter is invalid.

ErrNotFound

Indicates a requested resource does not exist.

ErrForbidden

Indicates an action is disallowed.

ErrInsufficientFunds

Indicates a balance of funds is insufficient.

ErrIllegalState

Indicates an actor’s internal state is invalid.

ErrSerialization

Indicates de/serialization failure within actor code.

ErrTooManyProveCommits

Power actor specific exit code.

ErrPlaceholder

Implementations

impl ExitCode

pub fn is_success(self) -> bool

returns true if the exit code was a success